service-workers/service-worker/fetch-event-is-reload-navigation-history.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
const worker = 'resources/fetch-event-test-worker.js';
const href = location.href.replace(/\?.*/, '');

promise_test(async (t) => {
  const scope = href;

  const reg = await service_worker_unregister_and_register(t, worker, scope);
  await wait_for_state(t, reg.installing, 'activated');
  const frame = await with_iframe(scope);
  assert_equals(frame.contentDocument.body.textContent,
                'method = GET, isReloadNavigation = false');
  await new Promise((resolve) => {
    frame.addEventListener('load', resolve);
    frame.contentDocument.body.innerText =
      'Reload this frame manually!';
  });
  assert_equals(frame.contentDocument.body.textContent,
      'method = GET, isReloadNavigation = true');
  frame.remove();
  await service_worker_unregister_and_done(t, scope);
}, 'Service Worker responds to fetch event with the correct isReloadNavigation value (manual reload)');

</script>
</body>
</html>
